// utils/generated/subgraph.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
}



export interface Block_Height {
  readonly hash?: Maybe<Scalars['Bytes']>;
  readonly number?: Maybe<Scalars['Int']>;
}

export interface Bundle {
  readonly __typename?: 'Bundle';
  readonly bnbPrice: Scalars['BigDecimal'];
  readonly id: Scalars['ID'];
}
// ...
export interface Factory {
  readonly __typename?: 'Factory';
  readonly id: Scalars['ID'];
  readonly pairCount: Scalars['BigInt'];
}
// ...
export interface Tvl {
  readonly __typename?: 'Tvl';
  readonly id: Scalars['ID'];
  readonly totalLiquidityUSD: Scalars['BigDecimal'];
}

export interface Pair {
  readonly __typename?: 'Pair';
  readonly id: Scalars['ID'];
  readonly reserve0: Scalars['BigDecimal'];
  readonly reserve1: Scalars['BigDecimal'];
  readonly reserveBNB: Scalars['BigDecimal'];
  readonly reserveUSD: Scalars['BigDecimal'];
  readonly token0: Token;
  readonly token0Price: Scalars['BigDecimal'];
  readonly token1: Token;
  readonly token1Price: Scalars['BigDecimal'];
  readonly trackedReserveBNB: Scalars['BigDecimal'];
  readonly trackedReserveUSD: Scalars['BigDecimal'];
  readonly untrackedVolumeBNB: Scalars['BigDecimal'];
  readonly untrackedVolumeUSD: Scalars['BigDecimal'];
  readonly volumeBNB: Scalars['BigDecimal'];
  readonly volumeToken0: Scalars['BigDecimal'];
  readonly volumeToken1: Scalars['BigDecimal'];
  readonly volumeUSD: Scalars['BigDecimal'];
}

export interface Transaction {
  readonly __typename?: 'Transaction';
  readonly id: Scalars['ID'];
}

export interface PairSimple {
  readonly __typename?: 'PairSimple';
  readonly address: Scalars['Bytes'];
  readonly id: Scalars['ID'];
  readonly token0: Token;
  readonly token1: Token;
}
// ...
export interface Query {
  readonly __typename?: 'Query';
  /** Access to subgraph metadata */
  readonly _meta?: Maybe<_Meta_>;
  readonly bundle?: Maybe<Bundle>;
  readonly bundles: ReadonlyArray<Bundle>;
  readonly factories: ReadonlyArray<Factory>;
  readonly factory?: Maybe<Factory>;
  readonly pair?: Maybe<Pair>;
  readonly pairSimple?: Maybe<PairSimple>;
  readonly pairSimples: ReadonlyArray<PairSimple>;
  readonly pairs: ReadonlyArray<Pair>;
  readonly token?: Maybe<Token>;
  readonly tokens: ReadonlyArray<Token>;
}
// ...
export interface Token {
  readonly __typename?: 'Token';
  readonly decimals: Scalars['BigInt'];
  readonly derivedBNB: Scalars['BigDecimal'];
  readonly derivedUSD: Scalars['BigDecimal'];
  readonly id: Scalars['ID'];
  readonly name: Scalars['String'];
  readonly symbol: Scalars['String'];
}
// ...
export interface _Meta_ {
  readonly __typename?: '_Meta_';
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   */
  readonly block: _Block_;
  /** The deployment ID */
  readonly deployment: Scalars['String'];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  readonly hasIndexingErrors: Scalars['Boolean'];
}

export enum _SubgraphErrorPolicy_ {
  /** Data will be returned even if the subgraph has indexing errors */
  Allow = 'allow',
  /** If the subgraph has indexing errors, data will be omitted. The default. */
  Deny = 'deny'
}

export type PairsVolumeQueryVariables = Exact<{
  limit: Scalars['Int'];
  pairIds: ReadonlyArray<Scalars['ID']> | Scalars['ID'];
  blockNumber: Scalars['Int'];
}>;


export type PairsVolumeQuery = (
  { readonly __typename?: 'Query' }
  & { readonly pairVolumes: ReadonlyArray<(
    { readonly __typename?: 'Pair' }
    & Pick<Pair, 'id' | 'volumeToken0' | 'volumeToken1'>
  )> }
);

export type TokenQueryVariables = Exact<{
  id: Scalars['ID'];
}>;

export type PairQueryVariables = Exact<{
  id: Scalars['ID'];
}>;


export type TokenQuery = (
  { readonly __typename?: 'Query' }
  & { readonly token?: Maybe<(
    { readonly __typename?: 'Token' }
    & Pick<Token, 'id' | 'name' | 'symbol' | 'derivedBNB' | 'derivedUSD'>
  )> }
);

export type PairQuery = (
  { readonly __typename?: 'Query' }
  & { readonly pair?: Maybe<(
    { readonly __typename?: 'Pair' }
  & Pick<Pair, 'id' | 'volumeToken0' | 'volumeToken1' | 'reserveUSD' | 'reserveBNB' | 'token0' | 'token1' | 'reserve0' | 'reserve1' | 'token0Price' | 'token1Price'>
    & { readonly token0: (
      { readonly __typename?: 'Token' }
      & TokenInfoFragment
    ), readonly token1: (
      { readonly __typename?: 'Token' }
      & TokenInfoFragment
    ) }
  )> }
);

// export type TvlQuery = (
//   { readonly __typename?: 'Query' }
//   & { readonly tvl?: Maybe<(
//     { readonly __typename?: 'Tvl' }
//     & Pick<Tvl, 'id' | 'totalLiquidityUSD' >
//   )> }
// );

export type TvlQuery = (
  { readonly __typename?: 'Query' }
  & { readonly tvl?: Maybe<(
    { readonly __typename?: 'Tvl' }
    & Pick<Tvl, 'id' | 'totalLiquidityUSD' >
  )> }
);

export type TokenInfoFragment = (
  { readonly __typename?: 'Token' }
  & Pick<Token, 'id' | 'name' | 'symbol' | 'derivedBNB' | 'derivedUSD'>
);

export type TopPairsQueryVariables = Exact<{
  limit: Scalars['Int'];
  excludeTokenIds: ReadonlyArray<Scalars['String']> | Scalars['String'];
}>;

export type TvlQueryVariables = Exact<{
  limit: Scalars['Int'];
}>;


export type TopPairsQuery = (
  { readonly __typename?: 'Query' }
  & { readonly pairs: ReadonlyArray<(
    { readonly __typename?: 'Pair' }
    & Pick<Pair, 'id' | 'reserve0' | 'reserve1' | 'volumeToken0' | 'volumeToken1' | 'reserveBNB' | 'reserveUSD'>
    & { readonly token0: (
      { readonly __typename?: 'Token' }
      & TokenInfoFragment
    ), readonly token1: (
      { readonly __typename?: 'Token' }
      & TokenInfoFragment
    ) }
  )> }
);

export type TopTransactionsQuery = (
  { readonly __typename?: 'Query' }
  & { readonly transactions: ReadonlyArray<(
    { readonly __typename?: 'Transaction' }
    & Pick<Transaction, 'id'>
  )> }
);